Implement zord movement

+++ src/game/totem.rs
@@ -1,5 +1,5 @@
 #[derive(Debug)]
 pub struct Totem {
-    pub x: i32,
-    pub y: i32,
+    pub x: i16,
+    pub y: i16,
 }
